ajout commentaires

+++ blog/views.py
@@ -9,17 +9,24 @@ from django.views.generic import CreateView
 from django.urls import reverse_lazy
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import Q
+from django.contrib import messages
+from django.core.exceptions import ValidationError
+from django.db.models import Count
+from django.views import View
+from django.core.exceptions import PermissionDenied
+
 
 
 
 @login_required
 def home(request):
-    # Liste des utilisateurs que l'utilisateur actuel suit
-    users_followed = models.UserFollows.objects.filter(user=request.user).values_list('followed_user', flat=True)
+    # users_followed = models.UserFollows.objects.filter(user=request.user).values_list('followed_user', flat=True)
 
-    tickets = models.Ticket.objects.filter(Q(user__in=users_followed) | Q(user=request.user))
-    reviews = models.Review.objects.filter(Q(user__in=users_followed) | Q(ticket__user=request.user))
+    # récupération de tous les tickets
+    tickets = models.Ticket.objects.all().annotate(review_count=Count('review'))
+    reviews = models.Review.objects.all()
 
+    # lambda est un mot-clé en Python qui vous permet de déclarer des fonctions anonymes
     posts = sorted(
         chain(tickets, reviews), 
         key=lambda post: post.time_created, 
@@ -29,15 +36,28 @@ def home(request):
     return render(request, 'blog/home.html', context={'posts': posts})
 
 
+
 #------------------------ticket--------------------
 
+"""
+LoginRequiredMixin : mixin qui s'assure que l'utilisateur est connecté 
+avant de permettre l'accès à la vue.
 
+CreateView est une vue générique basée sur une classe Django 
+pour créer de nouvelles instances d'un modèle.
+"""
 class TicketCreateView(LoginRequiredMixin, CreateView):
     model = models.Ticket
     form_class = forms.TicketForm
     template_name = 'blog/create_ticket.html'
     success_url = reverse_lazy('home')
 
+    """ 
+    avant de sauvegarder l'objet Ticket, 
+    on assigne l'utilisateur connecté actuellement  
+    à l'attribut user du Ticket. 
+    C'est pour cette raison qu'on surcharge la méthode form_valid.
+    """
     def form_valid(self, form):
         form.instance.user = self.request.user
         return super().form_valid(form)
@@ -45,16 +65,32 @@ class TicketCreateView(LoginRequiredMixin, CreateView):
 
 @login_required
 def edit_ticket(request, ticket_id):
+    
+    # récupération de l'instance du ticket par son id
     ticket = get_object_or_404(models.Ticket, id=ticket_id)
+    
+    # Vérifiez si l'utilisateur connecté est celui qui a créé le ticket
+    #  si utilisateur différent alors erreur 403 (interdiction)
+    if ticket.user != request.user:
+        raise PermissionDenied
+    
+    # Création d'un formulaire de modification prérempli avec les données du ticket existant
     edit_ticket = forms.TicketForm(instance=ticket)
+    
     delete_ticket = forms.DeleteTicketForm()
     
     if request.method == 'POST':
+        
         if 'edit_ticket' in request.POST:
+            """
+            Si le formulaire de modification est soumis, 
+            il est re-créé avec les nouvelles données entrées par l'utilisateur
+            """
             edit_ticket = forms.TicketForm(request.POST,request.FILES, instance=ticket)
             if edit_ticket.is_valid():
                 edit_ticket.save()
                 return redirect('home')
+            
         if 'delete_ticket' in request.POST:
             delete_ticket = forms.DeleteTicketForm(request.POST)
             if delete_ticket.is_valid():
@@ -62,8 +98,11 @@ def edit_ticket(request, ticket_id):
                 return redirect('home')
             
     else:
+        # formulaires vides
         edit_ticket = forms.TicketForm(instance=ticket)
         delete_ticket = forms.DeleteTicketForm()
+        
+    # formulaires sont passés au template pour être affichés
     context = {
         'edit_ticket': edit_ticket,
         'delete_ticket': delete_ticket,
@@ -71,7 +110,7 @@ def edit_ticket(request, ticket_id):
     return render(request, 'blog/edit_ticket.html', context=context)
 
 # ------------------------abonnement(subscribe)--------------
-from .forms import SearchUserForm
+
 
 @login_required
 def subscribe(request, user_id):
@@ -90,21 +129,24 @@ def subscribe(request, user_id):
     followers = models.UserFollows.objects.filter(followed_user=request.user)
 
     # Initialiser le formulaire de recherche
-    search_form = SearchUserForm(request.GET)
+    search_form = forms.SearchUserForm(request.GET)
     searched_users = None
 
     if search_form.is_valid():
         searched_username = search_form.cleaned_data["username"]
         # Créer une nouvelle liste d'utilisateurs qui correspondent à la recherche
         searched_users = all_users.filter(username__icontains=searched_username)
-        
 
     if request.method == "POST":
         user_to_follow = get_object_or_404(get_user_model(), id=request.POST.get('user_id'))
-        if user_to_follow not in followed_users:
+        try:
+            if user_to_follow == request.user:
+                raise ValidationError("Vous ne pouvez pas vous abonner à vous-même.")
+            if user_to_follow in followed_users:
+                raise ValidationError("Vous êtes déjà abonné à cet utilisateur.")
             models.UserFollows.objects.create(user=request.user, followed_user=user_to_follow)
-        else:
-            models.UserFollows.objects.filter(user=request.user, followed_user=user_to_follow).delete()
+        except ValidationError as e:
+            messages.error(request, e.message)
 
     context = {
         "user": user,
@@ -115,8 +157,10 @@ def subscribe(request, user_id):
         "search_form": search_form,
         "searched_users": searched_users 
     }
-    
+
     return render(request, "blog/subscribe.html", context=context)
+
+
 
 
 @login_required
@@ -179,7 +223,7 @@ def edit_review(request, review_id):
     return render(request, "blog/edit_review.html", context=context)
 
 
-from django.views import View
+# ----------------------ticket avec critique-------------------------------------------
 
 class CreateTicketAndReview(LoginRequiredMixin, View):
     template_name = 'blog/create_ticket_and_review.html'
@@ -204,3 +248,21 @@ class CreateTicketAndReview(LoginRequiredMixin, View):
 
             return redirect('home') 
         return render(request, self.template_name, {'form_ticket': form_ticket, 'form_review': form_review})
+    
+# ----------------------------------mes posts----------------------------------------------------
+
+@login_required
+def my_posts(request):
+    # Récupérer les Tickets et les Reviews de l'utilisateur connecté
+    tickets = models.Ticket.objects.filter(user=request.user).annotate(review_count=Count('review'))
+    reviews = models.Review.objects.filter(user=request.user)
+
+    posts = sorted(
+        chain(tickets, reviews),
+        key=lambda post: post.time_created,
+        reverse=True
+    )
+
+    return render(request, 'blog/my_posts.html', context={'posts': posts})
+
+
